image_to_3d/utils/model_generator.py
import os
import trimesh
import numpy as np
import matplotlib.pyplot as plt
from mpl_toolkits.mplot3d import Axes3D
import torch
from utils.image_processor import ImageProcessor
from utils.text_processor import TextProcessor

class Point2ImgToMesh:
    """Simple class that uses Point-E to generate meshes from images or text"""
    def __init__(self):
        # Would normally use Point-E or other text/image-to-3D models
        # For this prototype, we'll simulate the output with primitive shapes
        pass
        
    def generate_from_image(self, image):
        """Generate a 3D mesh from an image."""
        # In a real implementation, this would use Point-E or similar
        # For this prototype, we'll create a simple cube mesh
        vertices, faces = self._create_simple_mesh("cube")
        return vertices, faces
    
    def generate_from_text(self, text):
        """Generate a 3D mesh from text prompt."""
        # Determine shape type from text (very basic)
        shape_type = "sphere"
        if "car" in text.lower() or "vehicle" in text.lower():
            shape_type = "car"
        elif "cube" in text.lower() or "box" in text.lower():
            shape_type = "cube"
        elif "sphere" in text.lower() or "ball" in text.lower():
            shape_type = "sphere"
        elif "cylinder" in text.lower() or "tube" in text.lower():
            shape_type = "cylinder"
            
        vertices, faces = self._create_simple_mesh(shape_type)
        return vertices, faces
    
    def _create_simple_mesh(self, shape_type):
        """Create a simple mesh based on shape type."""
        if shape_type == "cube":
            # Create a cube
            vertices = np.array([
                [-1, -1, -1], [1, -1, -1], [1, 1, -1], [-1, 1, -1],
                [-1, -1, 1], [1, -1, 1], [1, 1, 1], [-1, 1, 1]
            ])
            faces = np.array([
                [0, 1, 2], [0, 2, 3],  # bottom
                [4, 5, 6], [4, 6, 7],  # top
                [0, 1, 5], [0, 5, 4],  # front
                [2, 3, 7], [2, 7, 6],  # back
                [0, 3, 7], [0, 7, 4],  # left
                [1, 2, 6], [1, 6, 5]   # right
            ])
        elif shape_type == "sphere":
            # Create a simple sphere approximation (icosphere)
            sphere = trimesh.creation.icosphere(subdivisions=2, radius=1.0)
            vertices = sphere.vertices
            faces = sphere.faces
        elif shape_type == "cylinder":
            # Create a cylinder
            cylinder = trimesh.creation.cylinder(radius=1.0, height=2.0)
            vertices = cylinder.vertices
            faces = cylinder.faces
        elif shape_type == "car":
            # Create a very basic car shape (cuboid with wheels)
            # Main body
            body = trimesh.creation.box(extents=[2.0, 1.0, 0.5])
            # Wheels (cylinders)
            wheel1 = trimesh.creation.cylinder(radius=0.25, height=0.2)
            wheel1.apply_translation([-0.5, -0.6, -0.25])
            wheel2 = trimesh.creation.cylinder(radius=0.25, height=0.2)
            wheel2.apply_translation([0.5, -0.6, -0.25])
            wheel3 = trimesh.creation.cylinder(radius=0.25, height=0.2)
            wheel3.apply_translation([-0.5, 0.6, -0.25])
            wheel4 = trimesh.creation.cylinder(radius=0.25, height=0.2)
            wheel4.apply_translation([0.5, 0.6, -0.25])
            
            # Combine all parts
            car = trimesh.util.concatenate([body, wheel1, wheel2, wheel3, wheel4])
            vertices = car.vertices
            faces = car.faces
        else:
            # Default to cube
            return self._create_simple_mesh("cube")
        
        return vertices, faces

class ModelGenerator:
    def __init__(self):
        self.mesh_generator = Point2ImgToMesh()
        
    def generate_from_image(self, image_path, output_path):
        """Generate 3D model from image and save to file."""
        # Process the image
        image_processor = ImageProcessor()
        processed_image = image_processor.preprocess_image(image_path)
        
        # Generate the mesh
        vertices, faces = self.mesh_generator.generate_from_image(processed_image)
        
        # Create and save the mesh
        mesh = trimesh.Trimesh(vertices=vertices, faces=faces)
        mesh.export(output_path)
        
        return mesh
    
    def generate_from_text(self, text_prompt, output_path):
        """Generate 3D model from text and save to file."""
        # Process the text
        text_processor = TextProcessor()
        processed_text = text_processor.process_text(text_prompt)
        
        # Generate the mesh
        vertices, faces = self.mesh_generator.generate_from_text(processed_text)
        
        # Create and save the mesh
        mesh = trimesh.Trimesh(vertices=vertices, faces=faces)
        mesh.export(output_path)
        
        return mesh
    
    def visualize_mesh(self, mesh):
        """Create a simple visualization of the mesh."""
        # Get vertices and faces
        vertices = mesh.vertices
        faces = mesh.faces
        
        # Create 3D plot
        fig = plt.figure(figsize=(10, 8))
        ax = fig.add_subplot(111, projection='3d')
        
        # Plot the mesh
        ax.plot_trisurf(vertices[:, 0], vertices[:, 1], vertices[:, 2],
                        triangles=faces, cmap='viridis', alpha=0.7)
        
        # Set labels and title
        ax.set_xlabel('X')
        ax.set_ylabel('Y')
        ax.set_zlabel('Z')
        ax.set_title('3D Model Visualization')
        
        # Equal aspect ratio
        ax.set_box_aspect([1, 1, 1])
        
        return fig